ByteBuffer -> ByteBuf

+++ communication-protocol-analysis/src/main/java/pub/guoxin/protocol/analysis/conf/cache/DataProtocolCache.java
@@ -9,12 +9,12 @@ import pub.guoxin.protocol.analysis.model.entity.DataProtocol;
  */
 public class DataProtocolCache {
 
-    private static ConcurrentHashMapCache<Short, DataProtocol> instance = new ConcurrentHashMapCache<>();
+    private static ConcurrentHashMapCache<Integer, DataProtocol> instance = new ConcurrentHashMapCache<>();
 
     private DataProtocolCache() {
     }
 
-    public static ConcurrentHashMapCache<Short, DataProtocol> getInstance() {
+    public static ConcurrentHashMapCache<Integer, DataProtocol> getInstance() {
         return instance;
     }
